refactor: extract diagnostic and recovery logic from ErrorBoundary.jsx
Resolves #9494 by extracting utility functions into errorDiagnostics.js.
